# internal/port/http/v1/openapi_type.gen.go
// Package v1 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package v1

// Defines values for DownloadLinkStatusResponseStatus.
const (
	DownloadLinkStatusResponseStatusError DownloadLinkStatusResponseStatus = "Error"
	DownloadLinkStatusResponseStatusReady DownloadLinkStatusResponseStatus = "Ready"
	DownloadLinkStatusResponseStatusWait  DownloadLinkStatusResponseStatus = "Wait"
)

// CreateDownloadRequest defines model for CreateDownloadRequest.
type CreateDownloadRequest struct {
	Links []string `json:"links"`
}

// DownloadLinkStatusResponse defines model for DownloadLinkStatusResponse.
type DownloadLinkStatusResponse struct {
	Link   string                           `json:"link"`
	Status DownloadLinkStatusResponseStatus `json:"status"`
}

// DownloadLinkStatusResponseStatus defines model for DownloadLinkStatusResponse.Status.
type DownloadLinkStatusResponseStatus string

// DownloadRequestResponse defines model for DownloadRequestResponse.
type DownloadRequestResponse struct {
	Responses []DownloadLinkStatusResponse `json:"responses"`
}

// Error defines model for Error.
type Error struct {
	Details string `json:"details"`
	Slug    string `json:"slug"`
}

// CreateRequestJSONRequestBody defines body for CreateRequest for application/json ContentType.
type CreateRequestJSONRequestBody = CreateDownloadRequest